Initial commit
Initial commit after playing aroud a little bit with React.
The plan is to quickly create most components and pages, then make any
necessary improvements, make the UI nicer with some css, and finally
integrate it with the backend.
Since this I'm using this to get more familiar with React, I'm not
using the usual TDD approach, but will rather setup tests later on.